Move rest of stuff to dyn_output.hpp
